19/05/23

+++ src/main/java/com/csidigital/rh/management/service/impl/ProjectServiceImpl.java
@@ -4,7 +4,9 @@ package com.csidigital.rh.management.service.impl;
 
 
 import com.csidigital.rh.dao.entity.ProjectReferenceSequence;
+import com.csidigital.rh.dao.entity.Resource;
 import com.csidigital.rh.dao.repository.ProjectReferenceSequenceRepository;
+import com.csidigital.rh.dao.repository.ResourceRepository;
 import com.csidigital.rh.management.service.ProjectService;
 
 
@@ -28,6 +30,8 @@ public class ProjectServiceImpl implements ProjectService {
     private ModelMapper modelMapper;
     @Autowired
     private  ProjectRepository projectRepository;
+    @Autowired
+    private ResourceRepository resourceRepository ;
     @Autowired
     private ProjectReferenceSequenceRepository sequenceRepository;
 
@@ -59,6 +63,14 @@ public class ProjectServiceImpl implements ProjectService {
     }
 
     @Override
+    public List<Resource> getProjectResource(Long id) {
+        Project project = projectRepository.findById(id)
+                .orElseThrow(()-> new ResourceNotFoundException("Project with id " +id+ " not found"));
+
+        return project.getResources();
+
+    }
+    @Override
     public ProjectDtoResponse createProject(ProjectDtoRequest projectDtoRequest) {
         ProjectReferenceSequence sequence = sequenceRepository.findById(1L)
                 .orElse(null) ;
@@ -69,8 +81,17 @@ public class ProjectServiceImpl implements ProjectService {
         Project project = modelMapper.map(projectDtoRequest, Project.class);
         String projectReference = String.format("PR_%04d", sequence.getNextValue());
         project.setProjectReference(projectReference);
+        List<Resource> existingResources = resourceRepository.findAllById(projectDtoRequest.getResourceIds());
+        for(Resource res : existingResources) {
+            res.setProject(project);
+        }
+        Resource responsable = resourceRepository.findById(projectDtoRequest.getResponsableNum()).orElseThrow();
+        project.setResources(existingResources);
+      //  project.setResponsable(responsable);
+        responsable.setProject(project);
+       Resource resp= resourceRepository.save(responsable);
+        project.setResponsable(resp);
         project = projectRepository.save(project);
-
         sequence.incrementNextValue();
         sequenceRepository.save(sequence);
 
@@ -83,13 +104,34 @@ public class ProjectServiceImpl implements ProjectService {
         Project project = projectRepository.findById(id)
                 .orElseThrow(()-> new ResourceNotFoundException("Project with id: " + id + " not found"));
         modelMapper.map(projectDtoRequest, project);
-      Project updatedProject = projectRepository.save(project);
+        List<Resource> existingResources = resourceRepository.findAllById(projectDtoRequest.getResourceIds());
+       Resource responsable = resourceRepository.findById(projectDtoRequest.getResponsableNum()).orElseThrow();
+        for(Resource res : existingResources) {
+            res.setProject(project);
+        }
+        project.setResponsable(responsable);
+        project.setResources(existingResources);
+        responsable.setProject(project);
+        resourceRepository.save(responsable);
+        Project updatedProject = projectRepository.save(project);
         return modelMapper.map(updatedProject, ProjectDtoResponse.class);
     }
 
     @Override
     public void deleteProjectById(Long id) {
-        projectRepository.deleteById(id);
+        Project project = projectRepository.findById(id).orElse(null);
+
+        if (project != null) {
+            // Remove the project reference from associated employees
+            List<Resource> resources = project.getResources();
+            for (Resource res : resources) {
+                res.setProject(null);
+            }
+
+            projectRepository.deleteById(id);
 
     }
 }
+
+}
+
